Rename imports to reflect current directories
